+++ LAB_ASSIGNMENT 7/PRACTICE PROGRAM/Q2/delete1.php
@@ -1,15 +1,21 @@
-<?php 
-$tn=$_GET["tn"];
+<?php
+$tn = $_GET["tn"];
+try {
+    $conn = mysqli_connect("localhost", "root", "", "php");
+    if ($conn) {
+        $query = "DELETE FROM ticket WHERE t_no='$tn'";
+        $result = mysqli_query($conn, $query);
+        if ($result) {
+            echo "<script>alert('Record Deleted!');</script>";
+            echo '<meta http-equiv="refresh" content="0;database.php">';
+        } else {
+            throw new Exception("Query Error:Record does not deleted!");
+        }
 
-$conn=mysqli_connect("localhost","root","","php");
-$query="DELETE FROM ticket WHERE t_no='$tn'";
-$result=mysqli_query($conn,$query);
-if($result)
-{
-    echo"<script>alert('Record Deleted!');</script>";
-    ?>
-    <meta http-equiv="refresh" content="0;database.php">
-    <?php
+    } else {
+        throw new Exception("Server Error:Record does not deleted!");
+
+    }
+} catch (Exception $error) {
+    echo "<script>alert('{$error->getMessage()}');</script>";
 }
-
-?>
